FastAPI/requirement.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.1.4
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.37.28
botocore==1.37.28
Brotli==1.1.0
catkin-pkg-modules==1.0.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
distro-info==1.7+build1
docutils==0.20.1
empy==3.3.4
exceptiongroup==1.2.2
fastapi==0.115.12
flake8==7.0.0
flake8-builtins==2.1.0
flake8-comprehensions==3.14.0
flake8-docstrings==1.6.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
fonttools==4.46.0
fs==2.4.16
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.12.0
incremental==22.10.0
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
lark==1.1.9
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.2.1
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
mpi4py==3.1.5
mpmath==0.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycodestyle==2.11.1
pycurl==7.45.3
pydantic==2.11.2
pydantic_core==2.33.1
pydocstyle==6.3.0
pydot==1.4.2
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
roman==3.3
rosdistro-modules==1.0.1
rospkg-modules==1.6.0
s3transfer==0.11.4
SciPy==1.11.4
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
starlette==0.46.1
sympy==1.12
systemd-python==235
Twisted==24.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
ufoLib2==0.16.0
unicodedata2==15.1.0
urllib3==2.0.7
uvicorn==0.34.0
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.15.0
zipp==1.0.0
zope.interface==6.1
